requirements.txt
enum34==1.1.6
horovod==0.13.5
matplotlib==2.2.2
numpy==1.14.2
opencv_python==3.4.3.18
photutils>=0.4
Pillow==5.3.0
pyquaternion==0.9.2
pyrr==0.9.2
PyYAML==3.13
rospkg==1.1.4
scipy==1.1.0
simplejson==3.16.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.25.0


# Exploratory
notebook==5.7.8
jupyter==1.0.0
git+https://github.com/EricCousineau-TRI/sixd_toolkit@dd610c4
pyassimp==4.1.4
pandas==0.24.2
